pyproject.toml
[tool.poetry]
name = "sketches"
version = "2024.01.01"
description = "Sketch a day, 2024 edition"
authors = ["Erico Andrei <ericof@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "utils"}, {include = "sketches"}]

[tool.poetry.dependencies]
python = "^3.11"
py5 = "^0.9.1a1"
CairoSVG = "^2.7.1"
opencv-python = "^4.8.1.78"
pymunk = "^6.6.0"
scikit-learn = "^1.3.2"
prettyconf = "^2.2.1"
click = "^8.1.7"
requests = "^2.31.0"
gitpython = "^3.1.40"

[tool.poetry.scripts]
sketches = 'utils.cli.cli:cli'

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"
pre-commit = "^3.6.0"

[tool.black]
target-version = ["py311"]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
